Restructure surfable code folders
