+++ include/surfable.h
@@ -1,0 +1,3 @@
+#ifndef GUARD_SURFABLE_H
+#define GUARD_SURFABLE_H
+#endif
