+++ ld_script.txt
@@ -344,7 +344,7 @@ SECTIONS {
         src/international_string_util.o(.text);
         src/pokemon_debug.o(.text);
         src/expansion_intro.o(.text);
-        src/surfable_pokemon.o(.text);
+        src/surfable.o(.text);
     } =0
 
     script_data :
@@ -712,7 +712,7 @@ SECTIONS {
         data/sound_data.o(.rodata);
         src/pokemon_debug.o(.rodata);
         src/expansion_intro.o(.rodata);
-        src/surfable_pokemon.o(.rodata);
+        src/surfable.o(.rodata);
     } =0
 
     song_data :
