+++ src/data/object_events/surfable/surfable_pokemon_templates.h
