+++ sym_ewram.txt
@@ -149,4 +149,4 @@
 	.include "src/trainer_hill.o"
 	.include "src/rayquaza_scene.o"
 	.include "src/debug.o"
-	.include "src/surfable_pokemon.o"
+	.include "src/surfable.o"
